+++ tests/test_auto_yahp.py
@@ -5,6 +5,7 @@ from typing import Dict, TextIO, Type, Union
 import pytest
 
 import yahp as hp
+from yahp.auto_hparams import generate_hparams_cls
 from yahp.types import JSON
 
 
@@ -48,17 +49,17 @@ class PrimitiveClassWithDefaults:
         self.int_arg = int_arg
 
 
-@pytest.mark.parameterize('use_default', [True, False])
+@pytest.mark.parametrize('use_default', [True, False])
 def test_primitive_class_with_defaults(use_default: bool):
     config = {}
     if not use_default:
         config['int_arg'] = 69
-    instance = hp.create(PrimitiveClass, config)
+    instance = hp.create(PrimitiveClassWithDefaults, config)
     assert isinstance(instance, PrimitiveClassWithDefaults)
     assert instance.int_arg == 42 if use_default else 69
 
 
-class NestedClassWithPrimitive:
+class ClassWithNestedNonHparams:
     """Class
 
     Args:
@@ -72,7 +73,12 @@ class NestedClassWithPrimitive:
         self.nested_arg = nested_arg
 
 
-class NestedClassWithHparams:
+class ClassWithNestedHparams:
+    """Class
+
+    Args:
+        nested_arg: Docstring
+    """
 
     def __init__(
         self,
@@ -81,56 +87,53 @@ class NestedClassWithHparams:
         self.nested_arg = nested_arg
 
 
-@pytest.mark.parametrize('cls', [NestedClassWithHparams, NestedClassWithPrimitive])
-def test_nested_class(cls: Union[Type[NestedClassWithHparams], Type[NestedClassWithPrimitive]]):
+@pytest.mark.parametrize('cls', [ClassWithNestedHparams, ClassWithNestedNonHparams])
+def test_class_with_nested_hparams(cls: Union[Type[ClassWithNestedHparams], Type[ClassWithNestedNonHparams]]):
     config: Dict[str, JSON] = {'nested_arg': {'int_arg': 42}}
     instance = hp.create(cls, config)
     assert isinstance(instance, cls)
     assert instance.nested_arg.int_arg == 42
 
 
-class UnsupportedTypeWithUnion:
+class UnionWithSupportedAndUnsupportedType:
     """Class
 
     Args:
-        file: Docstring
+        filename: Docstring
     """
 
     def __init__(
         self,
         # TextIO is not supported by yahp, but str is, so this should still work
-        file: Union[str, TextIO],
+        filename: Union[str, TextIO],
     ) -> None:
-        self.file = file
+        self.filename = filename
 
 
-def test_union_with_unsupported_types(use_default: bool):
-    config: Dict[str, JSON] = {'file': 'log.txt'}
-    instance = hp.create(UnsupportedTypeWithUnion, config)
-    assert isinstance(instance, UnsupportedTypeWithUnion)
-    assert instance.file == 'log.txt'
+def test_union_with_unsupported_types():
+    config: Dict[str, JSON] = {'filename': 'log.txt'}
+    instance = hp.create(UnionWithSupportedAndUnsupportedType, config)
+    assert isinstance(instance, UnionWithSupportedAndUnsupportedType)
+    assert instance.filename == 'log.txt'
 
 
 class UnsupportedType:
     """Class
 
     Args:
-        file: Docstring
+        filename: Docstring
     """
 
     def __init__(
         self,
         # YAHP cannot create a TextIO object from yaml
-        file: TextIO,
+        filename: TextIO,
     ) -> None:
-        self.file = file
+        self.filename = filename
 
 
 def test_unsupported_class_errors():
-    with pytest.raises(
-            NotImplementedError,
-            match=('UnsupportedType argument `file` of type `TextIO` cannot be parsed from a config map. To fix, '
-                   'use a Union with at least one supported type, or instead use an `hp.Hparams` dataclass.')):
+    with pytest.raises(TypeError, match=(r"Type annotation <class 'typing\.TextIO'> is not supported")):
         hp.create(UnsupportedType, {})
 
 
@@ -149,22 +152,43 @@ class ConcreteClass(AbstractClass):
         self.int_arg = int_arg
 
 
-class ClassWithAbstractArg:
+@dataclasses.dataclass
+class DataclassWithAbstractClass(hp.Hparams):
+    item: AbstractClass = hp.required('item')
+
+    hparams_registry = {'item': {'concrete': ConcreteClass,}}
+
+
+def test_dataclass_with_abstract_class():
+    config: Dict[str, JSON] = {'item': {'concrete': {'int_arg': 42}}}
+    instance = hp.create(DataclassWithAbstractClass, config)
+    assert isinstance(instance, DataclassWithAbstractClass)
+    assert isinstance(instance.item, ConcreteClass)
+    assert instance.item.int_arg == 42
+
+
+class ClassWithMixedHparamsRegistry:
     """Class
 
     Args:
-        cls: Docstring
+        cls (AbstractClass): Docstring.
     """
 
-    hparams_registry = {'cls': {'concrete': ConcreteClass,}}
+    hparams_registry = {
+        'cls': {
+            'concrete': ConcreteClass,
+            'auto_initialized': generate_hparams_cls(ConcreteClass),
+        }
+    }
 
     def __init__(self, cls: AbstractClass) -> None:
         self.cls = cls
 
 
-def test_unsupported_abstract_class_errors():
-    config: Dict[str, JSON] = {'cls': {'concrete': {'int_arg': 42}}}
-    instance = hp.create(ClassWithAbstractArg, config)
-    assert isinstance(instance, ClassWithAbstractArg)
+@pytest.mark.parametrize('key', ['concrete', 'auto_initialized'])
+def test_class_with_mixed_hparams_registry(key: str):
+    config: Dict[str, JSON] = {'cls': {key: {'int_arg': 42}}}
+    instance = hp.create(ClassWithMixedHparamsRegistry, config)
+    assert isinstance(instance, ClassWithMixedHparamsRegistry)
     assert isinstance(instance.cls, ConcreteClass)
     assert instance.cls.int_arg == 42
