+++ yahp/hparams.py
@@ -11,12 +11,11 @@ from abc import ABC
 from dataclasses import dataclass, fields
 from enum import Enum
 from io import StringIO
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, TextIO, Type, TypeVar, Union, cast, get_type_hints
+from typing import (TYPE_CHECKING, Any, Callable, Dict, List, Optional, TextIO, Type, TypeVar, Union, cast,
+                    get_type_hints)
 
 import yaml
 
-from yahp.create_object.commented_map import CMOptions, to_commented_map
-from yahp.create_object.create_from_hparams import create_from_hparams, get_argparse
 from yahp.utils import type_helpers
 from yahp.utils.iter_helpers import list_to_deduplicated_dict
 
@@ -54,7 +53,7 @@ class Hparams(ABC):
 
     # note: hparams_registry cannot be typed the normal way -- dataclass reads the type annotations
     # and would treat it like an instance variable. Instead, using the python2-style annotations
-    hparams_registry = {}  # type: Dict[str, Dict[str, Type["Hparams"]]]
+    hparams_registry = None  # type: Optional[Dict[str, Dict[str, Union[Callable[..., Any], Type["Hparams"]]]]]
 
     @classmethod
     def validate_keys(cls,
@@ -116,7 +115,8 @@ class Hparams(ABC):
         Returns:
             Hparams: An instance of the class.
         """
-        return create_from_hparams(cls, data=data, f=f, cli_args=cli_args)
+        from yahp.create_object.create_object import create
+        return create(cls, data=data, f=f, cli_args=cli_args)
 
     @classmethod
     def get_argparse(
@@ -125,6 +125,7 @@ class Hparams(ABC):
         data: Optional[Dict[str, JSON]] = None,
         cli_args: Union[List[str], bool] = True,
     ) -> argparse.ArgumentParser:
+        from yahp.create_object.create_object import get_argparse
         return get_argparse(cls, data=data, f=f, cli_args=cli_args)
 
     def to_yaml(self, **yaml_args: Any) -> str:
@@ -147,53 +148,55 @@ class Hparams(ABC):
         """
 
         res: Dict[str, JSON] = {}
-        field_types = get_type_hints(self.__class__)
         for f in fields(self):
             if not f.init:
                 continue
-            ftype = type_helpers.HparamsType(field_types[f.name])
             attr = getattr(self, f.name)
             if attr is None:  # first, take care of the optionals
                 res[f.name] = None
                 continue
-            if ftype.is_hparams_dataclass:
-                # Could be: List[Generic Hparams], Generic Hparams,
-                # List[Specific Hparams], or Specific Hparams
-                # If it's in the registry, it's generic. Otherwise, it's specific
-                if f.name in self.hparams_registry:
-                    inverted_registry = {v: k for (k, v) in self.hparams_registry[f.name].items()}
-                    if isinstance(attr, list):
-                        field_list: List[JSON] = []
-                        for x in attr:
-                            assert isinstance(x, Hparams)
-                            field_name = inverted_registry[type(x)]
-                            field_list.append({field_name: x.to_dict()})
-                        res[f.name] = list_to_deduplicated_dict(field_list)
-                    else:
-                        field_dict: Dict[str, JSON] = {}
-                        field_name = inverted_registry[type(attr)]
-                        # Generic hparams. Make sure to index by the key in the hparams registry
-                        field_dict[field_name] = attr.to_dict()
-                        res[f.name] = field_dict
-                else:
-                    # Specific -- either a list or not
-                    if isinstance(attr, list):
-                        res[f.name] = {str(i): x.to_dict() for i, x in enumerate(attr)}
-                    else:
-                        assert isinstance(attr, Hparams)
-                        res[f.name] = attr.to_dict()
-            else:
-                # Not a hparams type
+
+            # Is the field in the hparams registry?
+            if self.hparams_registry is not None and f.name in self.hparams_registry:
+                inverted_registry = {v: k for (k, v) in self.hparams_registry[f.name].items()}
                 if isinstance(attr, list):
-                    if ftype.is_enum:
-                        res[f.name] = [x.name for x in attr]
-                    else:
-                        res[f.name] = attr
+                    field_list: List[JSON] = []
+                    for x in attr:
+                        if not isinstance(x, Hparams):
+                            # Impossible to convert a class back into its hparams
+                            raise TypeError(
+                                f'Cannot serialize {type(self).__name__}.{f.name} of type {type(x).__name__}')
+                        field_name = inverted_registry[type(x)]
+                        field_list.append({field_name: x.to_dict()})
+                    res[f.name] = list_to_deduplicated_dict(field_list)
                 else:
-                    if isinstance(attr, Enum):
-                        res[f.name] = attr.name
+                    if not isinstance(attr, Hparams):
+                        # Impossible to convert a class back into its hparams
+                        raise TypeError(
+                            f'Cannot serialize {type(self).__name__}.{f.name} of type {type(attr).__name__}')
+                    field_dict: Dict[str, JSON] = {}
+                    field_name = inverted_registry[type(attr)]
+                    # Generic hparams. Make sure to index by the key in the hparams registry
+                    field_dict[field_name] = attr.to_dict()
+                    res[f.name] = field_dict
+
+            else:
+                # If it's not in the registry, then it must be specific
+                is_list = isinstance(attr, list)
+                if not is_list:
+                    attr = [attr]
+                ans = []
+                for x in attr:
+                    if isinstance(x, Hparams):
+                        ans.append(x.to_dict())
+                    elif x is None or isinstance(x, (str, float, bool, int, Enum, dict)):
+                        ans.append(x.name if isinstance(x, Enum) else x)
                     else:
-                        res[f.name] = attr
+                        raise TypeError(f'Cannot serialize {type(self).__name__}.{f.name} of type {type(x).__name__}')
+                if not is_list:
+                    ans = ans[0]
+                res[f.name] = ans
+
         return res
 
     def initialize_object(self, *args: Any, **kwargs: Any) -> Any:
@@ -232,8 +235,10 @@ class Hparams(ABC):
                 Whether to interactively generate the template.
                 Defaults to False.
         """
+        from yahp.create_object.commented_map import CMOptions, to_commented_map
+
         cm = to_commented_map(
-            cls=cls,
+            constructor=cls,
             options=CMOptions(
                 add_docs=add_docs,
                 typing_column=typing_column,
@@ -291,7 +296,7 @@ class Hparams(ABC):
             message = f'Unable to find field: {class_key}.{field} in: {cls.__name__}'
             logger.warning(message)
             raise ValueError(message)
-        if field not in cls.hparams_registry:
+        if cls.hparams_registry is None or field not in cls.hparams_registry:
             message = f'Unable to find field: {class_key}.{field} in: {cls.__name__} registry. \n'
             message += 'Is it a choose one or list Hparam?'
             logger.warning(message)
@@ -328,7 +333,7 @@ class Hparams(ABC):
                 if not isinstance(value, dict):
                     raise TypeError(f'{fname} must be a {ftype}; instead it is of type {type(value).__name__}')
                 continue
-            if not ftype.is_hparams_dataclass:
+            if ftype.is_primitive or ftype.is_enum:
                 if ftype.is_list:
                     if not isinstance(value, list):
                         raise TypeError(f'{fname} must be a {ftype}; instead it is of type {type(value).__name__}')
